consolidate shared components
