+++ ohmg/frontend/svelte_components/src/components/lists/Participants.svelte
@@ -2,7 +2,7 @@
   import { TableSort } from 'svelte-tablesort';
 
   import Link from '../base/Link.svelte';
-  import LoadingEllipsis from '../common/LoadingEllipsis.svelte';
+  import LoadingEllipsis from '../shared/LoadingEllipsis.svelte';
 
   import { getFromAPI } from '../../lib/requests';
