+++ ohmg/frontend/svelte_components/src/components/shared/LoadingEllipsis.svelte
