+++ ohmg/frontend/svelte_components/src/components/shared/buttons/InfoModalButton.svelte
@@ -1,7 +1,7 @@
 <script>
   import Question from 'phosphor-svelte/lib/Question';
 
-  import { getModal } from '../base/Modal.svelte';
+  import { getModal } from '../../base/Modal.svelte';
 
   export let modalId;
   export let size = '1em';
