+++ ohmg/frontend/svelte_components/src/components/shared/cards/MapItem.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Link from '../base/Link.svelte';
+  import Link from '../../base/Link.svelte';
 
   export let MAP;
 </script>
