+++ ohmg/frontend/svelte_components/src/components/shared/cards/NonMapCard.svelte
@@ -1,11 +1,9 @@
 <script>
   import MapTrifold from 'phosphor-svelte/lib/MapTrifold';
 
-  import { getModal } from '../base/Modal.svelte';
-
-  import Link from '../base/Link.svelte';
-
-  import BaseCard from '../base/Card.svelte';
+  import { getModal } from '../../base/Modal.svelte';
+  import Link from '../../base/Link.svelte';
+  import BaseCard from '../../base/Card.svelte';
 
   export let nonmap;
   export let userCanEdit;
