+++ ohmg/frontend/svelte_components/src/components/shared/modals/ConfirmUngeoreferenceModal.svelte
@@ -1,6 +1,6 @@
 <script>
-  import Modal, { getModal } from '../base/Modal.svelte';
-  import { submitPostRequest } from '../../lib/requests';
+  import Modal, { getModal } from '../../base/Modal.svelte';
+  import { submitPostRequest } from '../../../lib/requests';
 
   export let CONTEXT;
   export let layerId;
