+++ ohmg/frontend/svelte_components/src/components/shared/modals/GeoreferencedSectionModal.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Modal from '../base/Modal.svelte';
+  import Modal from '../../base/Modal.svelte';
   export let id;
 </script>
