+++ ohmg/frontend/svelte_components/src/components/shared/modals/ItemDownloadSectionModal.svelte
@@ -1,6 +1,6 @@
 <script>
-  import Modal from '../base/Modal.svelte';
-  import HowToUseXyz from '../overviews/sections/HowToUseXYZ.svelte';
+  import Modal from '../../base/Modal.svelte';
+  import HowToUseXyz from '../../overviews/sections/HowToUseXYZ.svelte';
 
   export let id;
 </script>
