+++ ohmg/frontend/svelte_components/src/components/shared/modals/MultiMaskModal.svelte
@@ -4,7 +4,7 @@
   import CropIcon from 'phosphor-svelte/lib/Crop';
   import Trash from 'phosphor-svelte/lib/Trash';
 
-  import Modal from '../base/Modal.svelte';
+  import Modal from '../../base/Modal.svelte';
   export let id;
 </script>
