+++ ohmg/frontend/svelte_components/src/components/shared/modals/ResourceDownloadSectionModal.svelte
@@ -1,0 +1,11 @@
+<script>
+  import Modal from '../../base/Modal.svelte';
+  import HowToUseXyz from '../../overviews/sections/HowToUseXYZ.svelte';
+
+  export let id;
+</script>
+
+<Modal {id}>
+  <h3>Download & Web Services</h3>
+  <HowToUseXyz />
+</Modal>
